% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tables.R
\name{gen_tables}
\alias{gen_tables}
\title{Fantasy function}
\usage{
gen_tables(year)
}
\arguments{
\item{year}{Simply specify the year to retrieve that year of stats in tabular format.}
}
\description{
This function scrapes the pro-football-reference website for the year you specify
}
\examples{
fantasy_2017 <- gen_tables(2017)
}
\keyword{cats}
